Final term feature

+++ app/src/main/java/com/example/bencanatracker/response/FloodResponse.kt
@@ -18,7 +18,7 @@ data class Results(
     val objects: Objectss,
 
     @SerializedName("arcs")
-    val arcs: List<List<List<Double>>>,
+    val arcs: List<List<List<Int>>>,
 
     @SerializedName("transform")
     val transform: Transform,
